+++ hash/api/lc_sha3.h
@@ -47,7 +47,7 @@ extern const struct lc_hash *lc_sha3_224;
 
 struct lc_sha3_224_state {
 	uint64_t state[LC_SHA3_STATE_WORDS];
-	size_t msg_len;
+	uint64_t msg_len;
 	size_t digestsize;
 	uint8_t offset;
 	uint8_t r;
@@ -95,7 +95,7 @@ extern const struct lc_hash *lc_sha3_256;
 
 struct lc_sha3_256_state {
 	uint64_t state[LC_SHA3_STATE_WORDS];
-	size_t msg_len;
+	uint64_t msg_len;
 	size_t digestsize;
 	uint8_t offset;
 	uint8_t r;
@@ -143,7 +143,7 @@ extern const struct lc_hash *lc_sha3_384;
 
 struct lc_sha3_384_state {
 	uint64_t state[LC_SHA3_STATE_WORDS];
-	size_t msg_len;
+	uint64_t msg_len;
 	size_t digestsize;
 	uint8_t offset;
 	uint8_t r;
@@ -191,7 +191,7 @@ extern const struct lc_hash *lc_sha3_512;
 
 struct lc_sha3_512_state {
 	uint64_t state[LC_SHA3_STATE_WORDS];
-	size_t msg_len;
+	uint64_t msg_len;
 	size_t digestsize;
 	uint8_t offset;
 	uint8_t r;
@@ -238,7 +238,7 @@ extern const struct lc_hash *lc_shake128;
 
 struct lc_shake_128_state {
 	uint64_t state[LC_SHA3_STATE_WORDS];
-	size_t msg_len;
+	uint64_t msg_len;
 	size_t digestsize;
 	uint8_t offset;
 	uint8_t r;
